Esc now supports int64
git-svn-id: svn://ultimatepp.org/upp/trunk@7686 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Esc/EscRun.cpp
@@ -100,7 +100,7 @@ int EscEscape::Int(int i)
 {
 	if(i >= arg.GetCount())
 		ThrowError("too little parameters" + InCall());
-	return esc.Int(arg[i], "parameter" + InCall());
+	return (int)esc.Int(arg[i], "parameter" + InCall());
 }
 
 void  EscEscape::CheckArray(int i)
